[TOPDOWN] Fix deconvolution bugs, improve noise-power estimation, drop DL code
FLASHDeconvAlgorithm:
- Fix unsigned underflow / OOB in MSn precursor intensity-sum loop
- Fix off-by-one (post-decrement) in FLASHIda-log CV lookup
- Avoid full-spectrum copy in mergeSpectra_ (use const ref)

SpectralDeconvolution:
- Fix unsigned-subtraction UB in harmonic-check bin offset (cast to int)

PeakGroup::getNoisePeakPower_:
- Fix spurious peak[0] addition in path-sum (check sentinel before add)
- break->continue in chain-start scan (edges are sparse, not a prefix)
- break->continue in divisor loop, restoring 2/3-isotope-skip
  connections; improves noise/harmonic rejection
- Perf (result-preserving): hoist per-call buffers, distance early-break
  in edge build O(N^2)->O(N*W), head-only path finding O(L^2)->O(L),
  thread_local scratch reuse

Remove unused/abandoned deep-learning scoring path: getDLVector,
getDLscore, DEEP_LEARNING blocks and DL-only constants/bindings.

+++ src/openms/source/ANALYSIS/TOPDOWN/SpectralDeconvolution.cpp
@@ -470,7 +470,7 @@ namespace OpenMS
                 int harmonic_cntr = 0;
                 if (ms_level_ > 1 && hc * abs_charge > current_max_charge_) { break; }
 
-                const int hdiff = (int)round((double)(next_iso_bin - mz_bin_index)) / hc * (hc / 2);
+                const int hdiff = (int)round((double)((int)next_iso_bin - (int)mz_bin_index)) / hc * (hc / 2);
                 const int next_harmonic_iso_bin = (int)mz_bin_index + hdiff;
                 // check if there are harmonic peaks between the current peak and the next isotope peak.
